requirements.txt
Authlib==1.0.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==36.0.2
Deprecated==1.2.13
Flask==2.0.3
Flask-MySQLdb==1.0.1
Flask-WTF==1.0.1
google-api-core==2.7.1
google-api-python-client==2.41.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
greenlet==1.1.2
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mysqlclient==2.1.0
oauthlib==3.2.0
packaging==21.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.7
redis==4.1.4
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
wrapt==1.13.3
WTForms==3.0.1
